Revert "ImageUser (#14)"
This reverts commit 6be46f61df0e2cd3e2bacfadd6d0825f08c3ec66.
