+++ Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/entidades/Offer.java
@@ -1,0 +1,4 @@
+package com.Egg.Inmobiliaria.entidades;
+
+public class Offer {
+}
