--- Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/entitys/ImageUser.java
@@ -1,62 +1,0 @@
-package com.Egg.Inmobiliaria.entitys;
-
-import javax.persistence.Basic;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Lob;
-import org.hibernate.annotations.GenericGenerator;
-
-@Entity
-public class ImageUser {
-    
-    @Id
-    @GeneratedValue(generator = "uuid")
-    @GenericGenerator(name = "uuid", strategy = "uuid2")
-    private String id;
-    
-    private String mime;
-    
-    private String nombre;
-    
-    @Lob @Basic(fetch = FetchType.LAZY)
-    private byte[] contenido;
-
-    public ImageUser() {
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getMime() {
-        return mime;
-    }
-
-    public void setMime(String mime) {
-        this.mime = mime;
-    }
-
-    public String getNombre() {
-        return nombre;
-    }
-
-    public void setNombre(String nombre) {
-        this.nombre = nombre;
-    }
-
-    public byte[] getContenido() {
-        return contenido;
-    }
-
-    public void setContenido(byte[] contenido) {
-        this.contenido = contenido;
-    }
-    
-    
-}
