+++ Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/enumeraciones/OfferStatus.java
@@ -1,4 +1,4 @@
-package com.Egg.Inmobiliaria.enums;
+package com.Egg.Inmobiliaria.enumeraciones;
 
 public enum OfferStatus {
     CLIENT_OFFER, ENTITY_REJECTED, ENTITY_ACCEPTED, INACTIVE_OFFER,CLIENT_ACCEPTED;
