+++ Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/enumeraciones/PropertyStatus.java
@@ -1,4 +1,4 @@
-package com.Egg.Inmobiliaria.enums;
+package com.Egg.Inmobiliaria.enumeraciones;
 
 public enum PropertyStatus {
     FOR_SALE, FOR_RENT,
