+++ Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/enumeraciones/PropertyType.java
@@ -1,4 +1,4 @@
-package com.Egg.Inmobiliaria.enums;
+package com.Egg.Inmobiliaria.enumeraciones;
 
 public enum PropertyType {
     CASA, DEPARTAMENTO, CABAÑA, DUPLEX,
