--- Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/repositories/ImageUserRepository.java
@@ -1,20 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package com.Egg.Inmobiliaria.repositories;
-
-import com.Egg.Inmobiliaria.entitys.ImageUser;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-/**
- *
- * @author rober
- */
-
-@Repository
-public interface ImageUserRepository extends JpaRepository<ImageUser, String>{
-    
-}
