+++ Inmobiliaria/src/main/resources/application.properties
@@ -1,8 +1,1 @@
-spring.datasource.url: jdbc:mysql://localhost:3306/inmobiliaria?allowPublicKeyRetrieval=true&useSSL=false&useTimezone=true&serverTimezone=GMT&characterEncoding=UTF-8
-spring.datasource.username: root
-spring.datasource.password: root
-spring.datasource.driver-class-name: com.mysql.cj.jdbc.Driver
-spring.jpa.show-sql: true
-spring.jpa.hibernate.ddl-auto: update
-spring.jpa.properties.hibernate.dialect: org.hibernate.dialect.MySQL5InnoDBDialect
-spring.thymeleaf.cache: false
+
